// asvolhClient/src/app/models/voluntarios.model.ts
export class Voluntarios {
    cedula?: string;
    nombres?: string;
    apellidos?: string;
    correo?: string;
    celular?: string;
    direccion?: string;
    tipo_usuario?: number;
    voluntariado?: number;
    fecha_nacimiento?: Date;
    fecha_ingreso?: Date;
    motivo_ingreso?: string;
}